Refactor loading
