+++ lib/loading/loading_page.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:logging/logging.dart';
-import 'package:nyrna/application/loading/loading.dart';
 
-import '../apps_list/apps_list.dart';
+import '../../apps_list/apps_list.dart';
+import 'loading.dart';
 
 /// Intermediate loading screen while verifying that Nyrna's dependencies are
 /// available. If they are not an error message is shown, preventing a crash.
