+++ lib/main.dart
@@ -9,10 +9,10 @@ import 'package:package_info_plus/package_info_plus.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:window_size/window_size.dart' as window;
 
+import 'app.dart';
 import 'apps_list/apps_list.dart';
 import 'infrastructure/app_version/app_version.dart';
 import 'logs/app_logger.dart';
-import 'presentation/app_widget.dart';
 import 'settings/cubit/settings_cubit.dart';
 import 'settings/settings_service.dart';
 import 'theme/theme.dart';
@@ -65,7 +65,7 @@ Future<void> main(List<String> args) async {
               appVersion: AppVersion(packageInfo),
             ),
             lazy: false,
-            child: const AppWidget(),
+            child: const App(),
           );
         },
       ),
